Replace `com.backbase.android.res.deferred` with `com.backbase.oss.deferredresources` (group) and `com.backbase.deferredresources` (package)
