+++ deferred-resources/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.backbase.android.res.deferred" />
+    package="com.backbase.deferredresources" />
