+++ deferred-resources/src/main/java/com/backbase/deferredresources/internal/Context.kt
@@ -1,0 +1,8 @@
+package com.backbase.deferredresources.internal
+
+import android.content.Context
+import androidx.core.os.ConfigurationCompat
+import java.util.Locale
+
+internal val Context.primaryLocale: Locale
+    get() = ConfigurationCompat.getLocales(resources.configuration)[0]
