+++ demo/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.backbase.android.res.deferred.demo">
+    package="com.backbase.deferredresources.demo">
 
   <application
       android:allowBackup="false"
